Customer_Email_Classifier+Response_Generator/requirements.txt
google-generativeai
pydantic
python-dotenv
